Решена проблема инвалидации ссылки owner в LineBorder
с помощью изменения vector на list в DataStorageManager

+++ ContoursDefiner/ContoursDefiner/ControlPointsCalculator.cpp
@@ -62,7 +62,7 @@ std::pair<int, int> GeneralBorderCalculator::calculateNearestPointsIdx(const Con
   result.first = firstIndex;
   result.second = secondNearPointIdx;
 
-  int countPassedPoints = 0;
+  size_t countPassedPoints = 0;
   while (dist > 5 && countPassedPoints < first.size())
   {
     int step = static_cast<int>(floor(dist));
